Split project into multiple files
Cloth, Target, and Water now exist as subclasses as the general
Obstacle; the first is ready for collision functionality to be added in,
while the target is awaiting a model and I still have to implement the
water. Also TODO are the camera improvements; The camera is now
contained in its own class, with a static instance called in the draw()
loop.